fixed leaderboards in all redsquare sidebars
