+++ mods/redsquare/lib/appspace/games.js
@@ -69,7 +69,7 @@ class RedSquareAppspaceGames {
         console.log(gameName);
         let gamemod = app.modules.returnModule(gameName);
         if (gamemod){
-          gamemod.overlay.show(app, mod, gamemod.returnGameRulesHTML());
+          gamemod.overlay.show(app, gamemod, gamemod.returnGameRulesHTML());
         }else{
           console.log("Module not found");
         }
