+++ mods/redsquare/lib/sidebar/sidebar.template.js
@@ -1,3 +1,4 @@
+const RedSquareLeagueSidebar = require('./league-sidebar.template');
 
 module.exports = RedSquareSidebarTemplate = (app, mod) => {
 
@@ -8,55 +9,7 @@ module.exports = RedSquareSidebarTemplate = (app, mod) => {
   <div class="redsquare-sidebar-calendar">
   </div>
 
-  <div class="saito-leaderboard">
-    <h6>Your Rankings:</h6>
-    <div class="saito-table">
-      <div class="saito-table-row odd">
-          <div class="saito-table-gamename">Twilight Struggle</div>
-          <div class="saito-table-rank">4</div>
-      </div>
-      <div class="saito-table-row">
-          <div class="saito-table-gamename">Red Imperium</div>
-          <div class="saito-table-rank">97</div>
-      </div>
-      <div class="saito-table-row odd">
-          <div class="saito-table-gamename">Spider Solitaire</div>
-          <div class="saito-table-rank">1435</div>
-      </div>
-      <div class="saito-table-row">
-          <div class="saito-table-gamename">Settlers of Saitoa</div>
-          <div class="saito-table-rank">352</div>
-      </div>
-      <div class="saito-table-row odd">
-          <div class="saito-table-gamename">Pandemic</div>
-          <div class="saito-table-rank">4242</div>
-      </div>
-      <div class="saito-table-row">
-          <div class="saito-table-gamename">Blackjack</div>
-          <div class="saito-table-rank">2222</div>
-      </div>
-      <div class="saito-table-row odd">
-          <div class="saito-table-gamename">Chess</div>
-          <div class="saito-table-rank">1283</div>
-      </div>
-      <div class="saito-table-row">
-          <div class="saito-table-gamename">Poker</div>
-          <div class="saito-table-rank">924</div>
-      </div>
-      <div class="saito-table-row odd">
-          <div class="saito-table-gamename">Solitrio Solitaire</div>
-          <div class="saito-table-rank">63</div>
-      </div>
-      <div class="saito-table-row">
-          <div class="saito-table-gamename">Wordblocks</div>
-          <div class="saito-table-rank">1452</div>
-      </div>
-      <div class="saito-table-row odd">
-          <div class="saito-table-gamename">Wuziqi</div>
-          <div class="saito-table-rank">843</div>
-      </div>
-  </div>
-
+  ${RedSquareLeagueSidebar(app, mod)}
 </div>
